requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==2.1.1
DateTime==5.2
Django==4.2.6
django-environ==0.11.2
dnspython==2.4.2
elementpath==4.1.5
future==0.18.3
idna==3.4
isodate==0.6.1
jsonschema==4.19.1
jsonschema-path==0.3.1
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
lxml==4.9.3
mongomock==4.1.2
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
packaging==23.2
pathable==0.4.3
psycopg2-binary==2.9.9
pyhandle==1.2.1
pymongo==3.12.3
PyMySQL==0.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
rdflib==7.0.0
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rpds-py==0.10.6
sentinels==1.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
validators==0.22.0
whitenoise==6.6.0
xmlschema==2.5.0
xmltodict==0.13.0
zope.interface==6.1
